RoPe and attention fp32.

+++ train_gpt2.py
@@ -201,10 +201,10 @@ class CausalSelfAttention(nn.Module):
         v = self.lambdas[0] * v + self.lambdas[1] * vi.view_as(v) # @KoszarskyB & @Grad62304977
         q, k = norm(q), norm(k) # QK norm @Grad62304977
 
-        q, k = self.rotary(q.float()).bfloat16(), self.rotary(k.float()).bfloat16()
-        y = flex_attention(q.transpose(1, 2), k.transpose(1, 2), v.transpose(1, 2), block_mask=block_mask)
+        q, k = self.rotary(q.float()), self.rotary(k.float())
+        y = flex_attention(q.transpose(1, 2), k.transpose(1, 2), v.transpose(1, 2).float(), block_mask=block_mask)
 
-        y = y.transpose(1, 2).contiguous().view_as(x)  # re-assemble all head outputs side by side
+        y = y.bfloat16().transpose(1, 2).contiguous().view_as(x)  # re-assemble all head outputs side by side
         y = self.c_proj(y)
         return y
